Fix refresh/access tokens

+++ backend/utils/auth/generateTokens.js
@@ -1,32 +1,41 @@
 const jwt = require('jsonwebtoken'); 
 const UserRefreshToken = require('../../models/UserRefreshToken');
 
-const generateTokens = async (user) => {
+const generateTokens = async (user, generateRefreshToken = true) => {
     try {
         const payload = { _id: user._id, roles: user.roles };
 
-        const accessTokenExp = Math.floor(Date.now() / 1000) + (1 * 10);
+        // Access token expires in 5 minutes
+        const accessTokenExp = Math.floor(Date.now() / 1000) + (5 * 60);
         const accessToken = jwt.sign(
-            { ...payload, exp: accessTokenExp},
+            { ...payload, exp: accessTokenExp },
             process.env.JWT_ACCESS_TOKEN_SECRET_KEY
-        )
+        );
 
-        const refreshTokenExp = Math.floor(Date.now() / 1000) + 60 * 60 * 24 * 5;
+        let refreshToken, refreshTokenExp;
 
-        const refreshToken = jwt.sign(
-            { ...payload, exp: refreshTokenExp },
-            process.env.JWT_REFRESH_TOKEN_SECRET_KEY
-        )
+        if (generateRefreshToken) {
+            // Refresh token expires in 5 days
+            refreshTokenExp = Math.floor(Date.now() / 1000) + (60 * 60);
+            refreshToken = jwt.sign(
+                { ...payload, exp: refreshTokenExp },
+                process.env.JWT_REFRESH_TOKEN_SECRET_KEY
+            );
 
-        await UserRefreshToken.findOneAndDelete({ userId: user._id })
+            // Replace the old refresh token with the new one
+            await UserRefreshToken.findOneAndDelete({ userId: user._id });
+            await new UserRefreshToken({ 
+                userId: user._id, 
+                token: refreshToken, 
+                expiresAt: new Date(refreshTokenExp * 1000) 
+            }).save();
+        }
 
-        await new UserRefreshToken({ userId: user._id, token: refreshToken }).save();
-
-        return Promise.resolve({ accessToken, refreshToken, accessTokenExp, refreshTokenExp });
+        return { accessToken, accessTokenExp, refreshToken, refreshTokenExp };
 
     } catch (error) {
         console.error('Error in generateTokens:', error);
-        return Promise.reject(error)
+        return Promise.reject(error);
     }
 }
